+++ src/screens/ManualBackupScreen/Card/Card.styles.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+	container: {
+		height: 40,
+		marginBottom: 16,
+		flexDirection: 'row',
+		width: '50%',
+		alignItems: 'center'
+	},
+	tag: {
+		height: 40,
+		width: 40,
+		borderRadius: 12,
+		justifyContent: 'center',
+		alignItems: 'center',
+		marginRight: 12
+	}
+});
