+++ src/screens/ManualBackupScreen/Card/Card.types.ts
@@ -1,0 +1,4 @@
+export interface CardProps {
+	title: string;
+	idx: number;
+}
