+++ src/screens/ManualBackupScreen/CopyButton/CopyButton.types.ts
@@ -1,0 +1,5 @@
+import { GestureResponderEvent } from 'react-native';
+
+export interface ButtonProps {
+	onPress?: (event: GestureResponderEvent) => void;
+}
