+++ src/screens/index.ts
@@ -16,7 +16,7 @@ import WalletCreatedScreen from './WalletCreatedScreen/WalletCreatedScreen';
 import WalletScreen from './WalletScreen/WalletScreen';
 import WelcomeScreen from './WelcomeScreen/WelcomeScreen';
 import DepositSuccessScreen from './DepositScreen/DepositSuccess/DepositSuccessScreen';
-import BackupToICloudScreen from './BackupToICloudScreen';
+import BackupToICloudScreen from './BackupToICloudScreen/BackupToICloudScreen';
 import TopUpWaitScreen from './TopUpWaitScreen/TopUpWaitScreen';
 import Test from './Test/Test';
